v1.4.4: Hack: account for lyrics width during formatting
The merge-in of the upstream v1.4.3 alongside with the downgrade of
Vexflow to v1.2.93 (to make it easier to follow the upstream) resulted
in regression: the note layout did not account for lyrics width, leading
to lyrics overlaps (sure, probably I'm doing something wrong in my
usecase). Thus, this bruteforce hack to ensure that minimum width of
each note is set according to its lyrics width. That works for my
usecase, but probably does not work as intended in many other
situations.